aggiornato cartella 7, lagrangiani

+++ source_7/algos/counting_sort.h
@@ -8,7 +8,7 @@
 #include <math.h>
 #include "../base/global.h"
 
-int counting_sort(int *input_edges, int *output_edges, int num_edges, double* weights, double wmin, double wmax);
+int counting_sort(int *input_edges, int *output_edges, int num_edges, int* weights, int wmin, int wmax);
 
 int counting_sort_2pass(int *input_edges, int *output_edges, int num_edges, int* weights, int wmin, int wmax, int block);
